Refactor dask_compute
